detangled list.h from all the detectors servers
