+++ slsDetectorServers/eigerDetectorServer/slsDetectorFunctionList.c
@@ -1,10 +1,10 @@
 // SPDX-License-Identifier: LGPL-3.0-or-other
 // Copyright (C) 2021 Contributors to the SLS Detector Package
+#include "slsDetectorFunctionList.h"
 #include "clogger.h"
 #include "common.h"
 #include "sharedMemory.h"
 #include "sls/versionAPI.h"
-#include "slsDetectorFunctionList_eiger.h"
 
 #ifndef VIRTUAL
 #include "Beb.h"
