+++ slsDetectorServers/xilinx_ctbDetectorServer/Makefile
@@ -22,7 +22,7 @@ PROGS		= 	xilinx_ctbDetectorServer
 DESTDIR		?= 	bin
 INSTMODE	= 	0777
 
-SRCS 		= 	slsDetectorFunctionList_xilinx.c  
+SRCS 		= 	slsDetectorFunctionList.c  
 SRCS		+=	$(main_src)slsDetectorServer.c $(main_src)slsDetectorServer_funcs.c $(main_src)communication_funcs.c $(main_src)arm64.c $(main_src)XILINX_PLL.c $(main_src)XILINX_FMC.c $(main_src)common.c $(main_src)/sharedMemory.c $(main_src)/loadPattern.c $(md5_dir)md5.c $(main_src)programViaArm.c $(main_src)LTC2620_Driver.c 
 
 OBJS 		= 	$(SRCS:.c=.o)
